Move projects page to app router
