+++ src/app/projects/page.tsx
@@ -1,11 +1,8 @@
-import { Container, Heading, VStack } from '@chakra-ui/react'
-import { subDays, format } from 'date-fns'
-import type { GetStaticProps } from 'next'
-import { NextSeo } from 'next-seo'
+import { format, subDays } from 'date-fns'
+import { type FC } from 'react'
 
-import ProjectCard from '~/components/ProjectCard'
-import type { GitHubRepo, Project } from '~/models'
-import { UserConfig } from '~/user.config'
+import ProjectsPage from '~/app/projects/ProjectsPage'
+import { type GitHubRepo, type Project } from '~/models'
 
 const PROJECTS_META_INFO: Array<{ githubRepo: string; packageUrl?: string }> = [
 	{
@@ -18,33 +15,11 @@ const PROJECTS_META_INFO: Array<{ githubRepo: string; packageUrl?: string }> = [
 ]
 
 const NPM_STAT_DATE_FORMAT = 'y-MM-dd'
-const REVALIDATE_SECONDS = 5
 
-interface ProjectsPageProps {
-	projects: Array<Project>
-}
-
-const ProjectsPage = ({ projects }: ProjectsPageProps) => {
-	return (
-		<>
-			<NextSeo
-				title="Projects"
-				description={`${UserConfig.author.name}'s Projects`}
-				openGraph={{ description: `${UserConfig.author.name}'s Projects` }}
-			/>
-			<Container maxWidth="container.md">
-				<Heading as="h1" variant="gradient" mb={4} fontSize="4xl">
-					Projects
-				</Heading>
-				<VStack spacing={6}>
-					{projects.map(({ repo, npmPackage }) => (
-						<ProjectCard key={repo.id} repo={repo} npmPackage={npmPackage} />
-					))}
-				</VStack>
-			</Container>
-		</>
-	)
-}
+// TODO: set metadata
+// title="Projects"
+// description={`${UserConfig.author.name}'s Projects`}
+// openGraph={{ description: `${UserConfig.author.name}'s Projects` }}
 
 function mapDataArrayToObjectCollection<DataType>(
 	arr: Array<{ url: string; data: DataType } | undefined>,
@@ -59,9 +34,7 @@ function mapDataArrayToObjectCollection<DataType>(
 	return collection
 }
 
-export const getStaticProps: GetStaticProps<{
-	projects: Array<Project>
-}> = async () => {
+async function getProjects(): Promise<Array<Project>> {
 	const GITHUB_ACCESS_TOKEN = process.env.GITHUB_ACCESS_TOKEN
 	const repos = await Promise.all(
 		PROJECTS_META_INFO.map(async ({ githubRepo }) => {
@@ -131,12 +104,14 @@ export const getStaticProps: GetStaticProps<{
 		}
 	}).filter(({ repo }) => !!repo)
 
-	return {
-		props: {
-			projects,
-		},
-		revalidate: REVALIDATE_SECONDS,
-	}
+	return projects
 }
 
-export default ProjectsPage
+export const revalidate = 5
+
+const Page: FC = async () => {
+	const projects = await getProjects()
+	return <ProjectsPage projects={projects} />
+}
+
+export default Page
